CLIC.Subcommand: split global and subcommand switch parsing
With this change, the global command and subcommand can use the same
switches and they won't clash.

One of the goal here is to realize the full potential of the
Switches_As_Args option where parsing of sub-command switches is
disabled.

For example this will allow:
$ alr build -f -XTEST=plop # -f is passed to gprbuild and not to alr
$ alr exec bash -c "echo lol" # -c is passed to bash and not to alr

This will break some command line that used to work, for instance:
$ alr with hello --non-interactive