Markdown help: fix problem with tables for options
The list of options got a line break after each option except for the
longest line. The reason is that two or more spaces at the end of a line
makes a hard line break in Markdown, but the longest line does not get
any padding at the right. This ensures the line break.